AI_Catalyst/requirements.txt
asyncpg>=0.28.0
aiohttp>=3.8.0
pyyaml>=6.0
psutil>=5.9.0
openai>=1.0.0
regex>=2023.0.0
pandas>=2.0.0